Split adapter implementation into single files
