+++ pymtl3/stdlib/mem/__init__.py
@@ -1,4 +1,4 @@
 from .MemoryFL import MemoryFL
 from .MemMsg import MemMsgType, mk_mem_msg, mk_mem_req_msg, mk_mem_resp_msg
 from .ROM import CombinationalROM, SequentialROM
-from .adapters import MemRequesterAdapterFL
+from .MemRequesterAdapterFL import MemRequesterAdapterFL
