+++ pymtl3/stdlib/stream/__init__.py
@@ -2,4 +2,5 @@ from .StreamSinkFL import StreamSinkFL
 from .StreamSourceFL import StreamSourceFL
 from . import ifcs
 from .queues import StreamNormalQueue, StreamPipeQueue, StreamBypassQueue
-from .adapters import IStreamDeqAdapterFL, OStreamEnqAdapterFL
+from .IStreamDeqAdapterFL import IStreamDeqAdapterFL
+from .OStreamEnqAdapterFL import OStreamEnqAdapterFL
