+++ pymtl3/stdlib/xcel/__init__.py
@@ -1,2 +1,2 @@
 from .XcelMsg import XcelMsgType, mk_xcel_msg, mk_xcel_req_msg, mk_xcel_resp_msg
-from .adapters import XcelRequesterAdapterFL
+from .XcelRequesterAdapterFL import XcelRequesterAdapterFL
